v3-dev-080719
v3 Development Release 3

This release adds colors to Watcher and Travis Status lists to better distinguish the results. There is also a traffic-light-like minimal UI which can be created for a selection of Watchers and Travis Statuses upon a click on the Overview-button. Please remove all existing Watchers and Travis Statuses before updating to this version to avoid possible errors.